api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.2.0
pyjwt>=2.10.0
cryptography>=44.0.0
httpx>=0.28.0
